Дописан контроллер Category, дописан тест для контроллера Category.

+++ app/controllers/category_controller.rb
@@ -6,25 +6,59 @@ class CategoryController < ApplicationController
     if @cat.save
       redirect_to @cat
     else
-      render :action =>:new
+      render :new
     end  
   end
   
   def show
-    @cat = Category.find(params[:id])
+    begin  
+      @cat = Category.find(params[:id])
+    rescue  
+#      redirect_to("#{Rails.root}/public/404.html") 
+#      render :file => "#{Rails.root}/public/404.html"
+      render :file => "#{Rails.root}/public/404.html", :layout => true, :status => 404
+#      render :file => "/home/alex/Desktop/ruby/ngkub/public/404.html", :layout => true, :status => 404
+    end 
   end
 
   def index
-    @cats = Category.all    
+    @cats = Category.all
+    if @cats.nil?
+      flash[:notice] = "No categories found"
+    end 
   end
 
   def delete
-    cat = Category.find(params[:category][:id])
-    cat.destroy    
+    begin  
+      cat = Category.find(params[:id])
+    rescue  
+      flash[:notice] = "we haven't category with id=" + params[:id]
+      render :file => "#{Rails.root}/public/404.html", :layout => true, :status => 404
+    end 
+    
+    if !(cat.nil?)
+      flash[:notice] = "category " + cat.title + " deleted"
+      cat.destroy
+      render :index  
+    end  
   end
 
   def edit
-    cat = Category.find(params[:category][:id])
+    begin  
+      @cat = Category.find(params[:id])
+    rescue  
+      render :file => "#{Rails.root}/public/404.html", :layout => true, :status => 404
+    end
+  end
+
+  def update
+    @cat = Category.find(params[:id])
+    
+    if @cat.update_attributes(params[:category])
+      redirect_to @cat
+    else
+      render :edit
+    end
   end
   
 end
